Add dump command to the Debugger.
To do that, Instructions class was updated with getProcedures() function and procedures array. Block class also received small updates, mainly related with handling the procedures array inherited form the Instructions class. The dump command itself was inplemented in the Debugger class, with the function dump() that check if the given file path is falid, and if it's valid, then dump() writes the current valuation and all declared procedures in the current scope.