Add Builder to ForLoop.
ForLoop now has a builder that works similar to the eg. StringBuilder. Builder is in the same file as the ForLoop class, so that ForLoop constructor can be private (non-accessible to the end-user), and the Builder doesn't need to implement any getters in order to later initialize the ForLoop constructor.