Add Builder to the IfStatement.
IfStatement now has a builder that works similar to the eg. StringBuilder. Builder is in the same file as the IfStatement class, so that IfStatement constructor can be private (non-accessible to the end-user), and the Builder doesn't need to implement any getters in order to later initialize the IfStatement constructor.