Add Builder to the Procedure.
Procedure now has a builder that works similar to the eg. StringBuilder. Builder is in the same file as the Procedure class, so that Procedure constructor can be private (non-accessible to the end-user), and the Builder doesn't need to implement any getters in order to later initialize the Procedure constructor.