Add static of() functio to all expressions in order to implement builder functionality.
of() function is public, and it takes all of the parameters that a given expression needs, and returns a new object of th egiven expression class. Now all expression classes have only a private constructor so that they can't be created in any other way than calling the of() function.