Add tests for builders.
As the title says, this commit adds tests for builders. Additionally, Debugger has received an overloaded ru() that takea additional runMode parameter to omit the question about the debug mode and other unnecessary prints. Procedure class will receive its tests in a different branch because it's a whole new construct in the language.