Add clone() function to all instructions in order to fix recursion on procedures.
clone() function makes a clone of a given instruction object by returning a new object of the instruction subclass. That object is created by using a copy constructor, that takes another object of the given Instructions subclass, and copies all its values to the new object. Using clone() function fixes the problem with the concurrent modification inside procedures in recursion.