Add procedure exceptions handling.
Exceptions related to procedures are now handled both in the Block and Procedure class, because both of these classes can have procedure declarations inside them. Error handling is realized by using booleans, which are set to true if the given procedure doesn't exist OR a map that holds declared procedures already contains a different procedure with the same name.